comment out log statement

+++ alphafold/data/pipeline_multimer.py
@@ -230,7 +230,7 @@ class DataPipeline:
             logging.info(
                 "Running monomer pipeline on chain %s: %s", chain_id, description
             )
-            logging.info("MSA_output_dir: %s", msa_output_dir)
+            # logging.info("MSA_output_dir: %s", msa_output_dir)
             chain_features = self._monomer_data_pipeline.process(
                 input_fasta_path=chain_fasta_path, msa_output_dir=chain_msa_output_dir
             )
@@ -275,12 +275,19 @@ class DataPipeline:
             sequences=input_seqs, descriptions=input_descs
         )
         chain_id_map_path = os.path.join(msa_output_dir, "chain_id_map.json")
-        with open(chain_id_map_path, "w") as f:
+        print(os.path.exists(chain_id_map_path))
+        if not os.path.exists(chain_id_map_path):
+            with open(chain_id_map_path, "w") as f:
+                chain_id_map_dict = {
+                    chain_id: dataclasses.asdict(fasta_chain)
+                    for chain_id, fasta_chain in chain_id_map.items()
+                }
+                json.dump(chain_id_map_dict, f, indent=4, sort_keys=True)
+        else:
             chain_id_map_dict = {
                 chain_id: dataclasses.asdict(fasta_chain)
                 for chain_id, fasta_chain in chain_id_map.items()
             }
-            json.dump(chain_id_map_dict, f, indent=4, sort_keys=True)
 
         all_chain_features = {}
         sequence_features = {}
